Refactor scaffolding script
We can simplify the script to only focus on journey logic.

This change removes necessary templates we assumed would be useful, but we do not currently have a need for them.
