--- templates/fetch.test.js
@@ -1,30 +1,0 @@
-'use strict'
-
-// Test framework dependencies
-const Lab = require('@hapi/lab')
-const Code = require('@hapi/code')
-const Sinon = require('sinon')
-
-const { describe, it, afterEach, beforeEach } = (exports.lab = Lab.script())
-const { expect } = Code
-
-// Thing under test
-const __MODULE_NAME__ = require('__REQUIRE_PATH__')
-
-describe('__DESCRIBE_LABEL__', () => {
-  let session
-
-  beforeEach(async () => {})
-
-  afterEach(() => {
-    Sinon.restore()
-  })
-
-  describe('when called', () => {
-    it('returns page data for the view', async () => {
-      const result = await __MODULE_NAME__.go(session.id)
-
-      expect(result).to.equal({})
-    })
-  })
-})
