+++ templates/presenter.test.js
@@ -13,11 +13,11 @@ const __MODULE_NAME__ = require('__REQUIRE_PATH__')
 describe('__DESCRIBE_LABEL__', () => {
   let session
 
-  beforeEach(async () => {})
+  beforeEach(() => {})
 
   describe('when called', () => {
     it('returns page data for the view', () => {
-      const result = __MODULE_NAME__.go(session.id)
+      const result = __MODULE_NAME__.go(session)
 
       expect(result).to.equal({})
     })
